[DEMO-2561] Remove datastore, replace test for FileInventoryRepository
* Since we no longer use / need the datastore class, remove it. Replace
the test written for it to use FileInventoryRepository instead.
